Add AppendingTextBox
Custom TextBox implementation binding to an ItemSource instead of text.
When items are added, call AppendText method to add to current text
content. Doing so will not replace text, so any current selections or
other text-dependent state will not be altered when new text is added.

Control exposes scroll behavior, allowing for always scrolling to the
new appended text, only scrolling if view is already at the bottom, or
using default behavior.

This control is intended to be used as a tail following a view of text
that is being updated in real time, such as displaying a running log.