Add executor to abstract running Python
Abstraction allows Python to be executed in different forms while always
being surrounded in common boilerplate code.

Support an interface for custom reusable Python middleware that can
modify the Python scope state both before and after the delegated
execution, and cleanup hooks for finalization. This can be used if
common code needs to be set up regardless of what Python executor
implementation is going to be used.