Add executor to run a function wrapped script
Wrapping a script in a function allows for injecting parameters to be
available to the script. It also could allow for adding decorators.
It is abstracted such that more arguments can easily be added.

This is not the cleanest implementation, but it was recognized that any
script needs to be wrapped in a catch for KeyboardInterrupt to handle
the cancellation token PythonEngine.Interrupt. Not doing so causes an
issue where the KeyboardInterrupt exception is then attempted to be
wrapped in a CLR exception but the Python thread is missing, causing
a NullReferenceException if the task is cancelled. More work can be done
to clean up the wrapper implementation.