Add redirect of Python stdout to output view
Redirect sys.stdout and sys.stderr via a callback that publishes
to a queue that is then rendered in the main application.

This allows scripts to view output via print statements and any
errors directly in an output tab on the right.

Use custom appending text box control for intuitive user experience.