Add matplotlib renderer
Searches pyplot state on script post-processing and renders a
non-interactive image from any figures. Figures are added as separate
output tabs and reset when script is run again.

Figures are pickled for future re-render. On size changed or tab load
state changed, figures is re-rendered at the appropriate dimensions.
This allows for smooth resizing of the image in the application.