Convert to data filter node
Making the base analysis node into a filter type node allows for
child analyses or filters to be added. The node should now register
a method that will control the filtered data. Data filters by default
return transformed data from GetIonData method, so create a new
GetOwnerData method that matches the previous functionality of
GetIonData when this was an analysis type node. Update all uses of
GetIonData to instead use GetOwnerIonData so that no functionality
is changed.