Fix installation issues
Target lower version of framework.

AP Suite targets net6.0-windows with OSPlatform 10.0
This makes the version for package resolution effectively
net6.0-windows10.0. However this is not a valid TargetFramework,
and the lowest for version 10 is 10.0.17763.0. However that is still
higher than the effective 10.0.0.0, causing nothing to be installed.

Could be solved by setting OSPlatform attribute using AssemblyInfo,
but that takes more work and we actually don't care about OS version
specifically for this extension, so easier solution is so just pick a
lower OS version than 10.0.0.0 that will then be installed.