Update nuspect to move hierarchical data to content
Due to a bug in AP Suite 6.3.0, lib content subdirectory folder structure
is not preserved on extension installation. However, content structure is.
In this case, as we don't need to multi-target frameworks and only support
the single platform for now, simply putting files with relevant structure
into content will allow this extention to be properly installed in 6.3.0